refactored settings layout
This is a bit unconventional, but what I wanted to do was make it so
that there was a diretory for every target host.  It doesn't seem very
sensible with just the settings.py in the picture, but it makes a lot
more sense when you start throwing other services into the mix.
Otherwise there would have been nginx/prod.conf nginx/dev.conf
nginx/local.conf instead of local/nginx.conf, dev/nginx.conf, etc.  This
makes it a lot easier to do things like omit from version control the
information for a single host (e.g., local).